requirements.txt
Django~=3.1.5
djangorestframework~=3.12.2
djangorestframework-jwt~=1.11.0
django-cors-headers~=3.7.0
django-filter~=2.4.0
reportlab~=3.5.59
pandas~=1.2.2
requests~=2.25.1
PyJWT~=1.7.1
pytz~=2021.1